Implemented Rendering for All Currently Implemented Blocks/Items Rendering for the glass button and glass pressure plate was broken in ICBM-Classic 1.12.2, this implementation approximately matches the original ICBM 1.5.2 version used in Voltz but imo it still looks like garbage and should probably be redone at some point.
Also deleted the textures/block/reinforced_glass folder which was presumably intended for use with connected textures, wasn't implemented in 1.12.2 and not worth the time to write it from scratch.

Also the CreativeTab for glass buttons and glass pressure plates was CreativeTabs.REDSTONE in 1.12.2, switching both to the dedicated ICBM ItemGroup in this version because imo mixing modded items into the vanilla groups when a custom one exists is counterintuitive and sloppy.