Entity Registration and Spawning for Explosives Currently just an unskinned clone of Primed TNT and only implemented for Condensed Explosives (could easily be applied to the others but no point until it actually does something)
Moved Condensed Explosives to a dedicated class to test the entity spawning, both the blocks and the entities will need custom base classes combined with minimal code to tweak each explosive in the registration classes (current plan is to implement a dedicated event for each type of blast which the registration class would just need to pass a factory of to the base class' constructor. This would have the added effect of enabling other mods to cancel/intercept blasts).