Generalized Multiblock BlockState Properties and Placement Routine Still in the BlockLauncherPlatform class for now, will separate once ready to apply to other blocks.
Replace "half" with three separate properties for horizontal/height/depth coordinates and three others to flag each as being negative (IntegerProperty does not support negative values, so adding a BooleanProperty for each is a workaround).

The new properties required a new placement routine - took the opportunity to generalize it an arbitrary shape given by an array Vector3i coordinates.