Basic Ore Generation, Fix BlastEvent Client Routine Ore gen is simple, vanilla-style with zero attempt made at balancing. Need to balance, add config, and ideally implement custom generation for sulfur to spawn around lava pools.
BlastEvent previously failed to play sounds or spawn particles on the client side. Had the start of some code to synchronize the event across the network, opted instead to embrace the server-side-only design and piggyback on some methods from ServerWorld to handle the networking. See com.jdawg3636.icbm.common.event.ICBMEvents for implementation.