Add Mixins, Rocket Launcher Render, Rename "blocks" Package to "block"
Added rocket launcher model (.obj model derived with slight modification from the MC 1.5.2 version, .json file created manually with the goal of mimicking the best of the old versions while improving upon their weaknesses).

To augment the launcher model, there is now a non-required (i.e. will not crash the game if it fails) Mixin that transforms net.minecraft.client.renderer.entity.PlayerRenderer such that the third-person player model holds the launcher up on its shoulder instead of down by its side.

A new "item" package has been created for items that require custom classes (the rocket launcher has one but it currently contains no unique functionality; it was created prior to resorting to a Mixin for the third-person rendering tweak). To match the singular-form naming, the "blocks" package has been renamed to "block". This is more consistent with the convention used elsewhere in Forge Modding and in the Vanilla Game.