Missile Flight Logic (First Draft)
Flight is currently staccato, needs some interpolation on the client side. Also need to implement rotation for particle emission and rendering, fix trajectory generation to take into account altitude, etc.

Demo Command:
/summon icbm:missile_incendiary ~ ~ ~ {MissileLaunchPhase:2,SourcePosX:8280,SourcePosY:70,SourcePosZ:-1656,DestPosX:8325,DestPosY:70,DestPosZ:-1656,PeakHeight:80F,TotalFlightTicks:100}

NOTE ON THE DEMO COMMAND: You must spawn the missile at the same location that you specify for the SourcePosX/Y/Z NBT parameters in order to avoid extremely glitchy behavior. Also note that the Y values are currently ignored for both the source and destination parameters so the missile will always spawn at y=0. The missile will explode if it hits bedrock, so this means that there needs to be exposed void surrounding the source for the missile to not immediately explode on launch.