Overhaul Blast Event Dispatching
Moved every blast into its own class with an executeBlast() method which is called from AbstractBlastEvent.fire(). This fire() method checks that the event has not been cancelled prior to executing a blast, which the previous system did not do. The separate classes also result in much clearer code and package structure.