Improve Camouflage Lighting
One remaining issue which I'm unlikely to fix: light sources on one side of a camo block will illuminate blocks on the other. This is because we must use the noOcclusion BlockState property to make sure that the camo block doesn't stop other blocks' faces from rendering (this is even true if the camo appearance is set to an opaque cube, as it could still cause an X-Ray effect if the player steps inside of it). This same flag is used for lighting calculations. Thus, a Mixin would be required to target one of the two system and allow for differentiated behavior. Particularly for 1.16.5, this is not important enough to justify the invasiveness or development time.