Fix Missile Data Staying Tied to Creative Player Inv After Placing
When a platform was connected to a liquid fuel source, I was previously observing that the missile item in the hotbar would become fueled after right-clicking to place it on the platform, even though only the COPY of the missile that is placed should have the fuel added. This issue was creative-mode only as the ItemStack is removed from the player's inventory otherwise.